Fix dependency-groups (#95)
The original implementation of dependency-groups for the CLI treats
rules as a list of strings.
The actual implementation of dependency-groups is a hash where the
key is the type of dependency-group and the value is the group's rules.

For now there is only one type of dependency-group that is supported,
"patterns". However, in the future there may be more types that are
supported, so we must specify the type here.

This PR updates dependency-groups so that rules are now of type
`map[string]any`. The job test reflects this change by passing in
a key/value pair of rules in the form of `pattern: [patterns]`