The `package-managers` field is now optional
The reason I switched us from `record_package_manager_version` to
`record_ecosystem_versions` is that we want to start recording the
language runtime version as well.

There may be situations where we have instrumented the language but not
the package manager version.

In those cases, this field will be empty, so make it optional.