Reset `os.Stdin` after test
This test munges `os.Stdin`, which contains global state, so reset it
after the test is over to what it was before the test ran.

Also a little minor tidying up of `t.Cleanup()` to ensure all the
cleanups get registered on the same `defer` stack.

And check errors whenever a method may emit them to ensure no other
unexpected surprises.

This issue btw was flushed out by trying to run the tests with
`-count=2`.